# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fitnessstracker server.",
        "title": "Fitnesstracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/user/exercise/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update exercise details like Duration, ExerciseType, CaloriesBurned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "update exercise details of today",
                "operationId": "user-exercise-update",
                "parameters": [
                    {
                        "description": "The input for update the exercise details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "add new exercise details with ExerciseType, Duration, CaloriesBurned, Date(default)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "Add a new exercise for today",
                "operationId": "user-exercise-add",
                "parameters": [
                    {
                        "description": "The input for add exercise",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "delete exercise details of today with ExerciseType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "delete exercise details of today",
                "operationId": "user-exercise-delete",
                "parameters": [
                    {
                        "description": "The input for delete exercise",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExerciseType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/meal/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update meal details like Ingredients, MealType, CaloriesConsumed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "update meal details of today",
                "operationId": "user-meal-update",
                "parameters": [
                    {
                        "description": "The input for update the meal details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "add new meal details with MealType, Ingredients, CaloriesConsumed, Date(default)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "Add a new meal details for today",
                "operationId": "user-meal-add",
                "parameters": [
                    {
                        "description": "The input for add meal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "delete meal details of today with MealType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "delete meal details of today",
                "operationId": "user-meal-delete",
                "parameters": [
                    {
                        "description": "The input for delete meal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MealType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
        "/user/water/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update daily water details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Water"
                ],
                "summary": "update water details of today",
                "operationId": "user-water-update",
                "parameters": [
                    {
                        "description": "The input for update the water details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Water"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "add water details with in WaterIntake(in litre), Date(default)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Water"
                ],
                "summary": "Add water consumption for today",
                "operationId": "user-water-add",
                "parameters": [
                    {
                        "description": "The input for add daily water",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Water"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "delete water details of today",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Water"
                ],
                "summary": "delete water details of today",
                "operationId": "user-water-delete",
                "responses": {
                    "200": {
                        "description": "User details Successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Exercise": {
            "type": "object",
            "required": [
                "caloriesBurned",
                "duration",
                "exerciseType"
            ],
            "properties": {
                "caloriesBurned": {
                    "type": "number",
                    "maximum": 20000
                },
                "date": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer",
                    "maximum": 1440
                },
                "exerciseType": {
                    "type": "string",
                    "enum": [
                        "weight_lifting",
                        "walking",
                        "running",
                        "gym",
                        "yoga"
                    ]
                }
            }
        },
        "models.ExerciseType": {
            "type": "object",
            "required": [
                "exerciseType"
            ],
            "properties": {
                "exerciseType": {
                    "type": "string",
                    "enum": [
                        "weight_lifting",
                        "walking",
                        "running",
                        "gym",
                        "yoga"
                    ]
                }
            }
        },
...
        "models.Meal": {
            "type": "object",
            "required": [
                "caloriesConsumed",
                "ingredients",
                "mealType"
            ],
            "properties": {
                "caloriesConsumed": {
                    "type": "number",
                    "maximum": 20000,
                    "minimum": 0
                },
                "date": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "string"
                },
                "mealType": {
                    "type": "string",
                    "enum": [
                        "breakfast",
                        "lunch",
                        "snacks",
                        "dinner"
                    ]
                }
            }
        },
        "models.MealType": {
            "type": "object",
            "required": [
                "mealType"
            ],
            "properties": {
                "mealType": {
                    "type": "string",
                    "enum": [
                        "breakfast",
                        "lunch",
                        "snacks",
                        "dinner"
                    ]
                }
            }
        },
        "models.Message": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Water": {
            "type": "object",
            "required": [
                "waterIntake"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "waterIntake": {
                    "type": "number",
                    "maximum": 20
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}